+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## [Unreleased]
+
+### Added
+
+- github: new action `forkRepo`.
+
+
+
 ## [1.6.0] - 2019-04-03
 
 ### Added